Rewrote jsxhint.
JSXHint should maintain parity with JSHint for a minimum of surprises
when using with tools that expect JSHint.

This new version transforms input files and streams, writes to temporary files,
lints those using the same input options as were passed to jshint, and
translates output so the transformation is transparent.

All jshint flags should work with jsxhint exactly as they work with jshint.