Treat JSX transpiler errors like JSHint errors
Prior to this commit, any error that occurred within the JSX transpiler would
crash the process. Obviously this is not very helpful as all linting is thrown
out the window. This commit captures any error thrown in the transpile step and
transfers it to the JSHint reporter so it will be displayed like any other
linter error.