Create temporary files in os.tmpdir()
Creating temporary files in the working tree can cause havok when it comes to
file watchers. This commit will save all temp files in the directory specified
by `os.tmpdir()`, so it *should* be safe on Windows too. It also ensures to copy
over the parent `.jshintrc` and also `package.json` as JSHint CLI uses these for
options.