Strip Flow type annotations when transforming
Facebook has recently released Flow, a static type checker for
JavaScript.

  http://flowtype.org/

Flow augments the standard JavaScript syntax with type annotations that
need to be removed before passing the code on to things that expect
actual JavaScript. The react-tools project includes a stripTypes option,
that we can just turn on to process this code correctly.

This commit allows people to use Flow with jsxhint by including the
stripTypes option when performing the react-tools transformation.

Closes #35